fix: refactor user error & exception message gf-6
