--- apps/backend/src/libs/exceptions/user-error/user-error.exception.ts
@@ -1,9 +1,0 @@
-import { ApplicationError } from "@git-fit/shared";
-
-class UserError extends ApplicationError {
-	public constructor(message: string) {
-		super({ message });
-	}
-}
-
-export { UserError };
