+++ apps/backend/src/modules/users/libs/enums/enums.ts
@@ -1,2 +1,1 @@
-export { ExceptionMessage } from "./exception-message.enum.js";
-export { UsersApiPath } from "@git-fit/shared";
+export { ExceptionMessage, UsersApiPath } from "@git-fit/shared";
