+++ packages/shared/src/libs/modules/http/libs/enums/http-code.enum.ts
@@ -1,4 +1,5 @@
 const HTTPCode = {
+	CONFLICT: 409,
 	CREATED: 201,
 	INTERNAL_SERVER_ERROR: 500,
 	OK: 200,
