+++ packages/shared/src/modules/users/libs/exceptions/exceptions.ts
@@ -1,0 +1,1 @@
+export { UserError } from "./user-error.exception.js";
