+++ packages/shared/src/modules/users/libs/exceptions/user-error.exception.ts
@@ -1,0 +1,24 @@
+import { type ValueOf } from "src/libs/types/types.js";
+
+import {
+	type HTTPCode,
+	HTTPError,
+} from "../../../../libs/modules/http/http.js";
+
+type Constructor = {
+	cause?: unknown;
+	message: string;
+	status: ValueOf<typeof HTTPCode>;
+};
+
+class UserError extends HTTPError {
+	public constructor({ cause, message, status }: Constructor) {
+		super({
+			cause,
+			message,
+			status,
+		});
+	}
+}
+
+export { UserError };
