+++ packages/shared/src/modules/users/users.ts
@@ -1,4 +1,5 @@
 export { UsersApiPath } from "./libs/enums/enums.js";
+export { UserError } from "./libs/exceptions/exceptions.js";
 export {
 	type UserGetAllItemResponseDto,
 	type UserGetAllResponseDto,
